+++ Assets/RicKit/UI/CHANGELOG.md
@@ -1,4 +1,7 @@
-# Changelog
+# Changelog‘
+## [2.0.0] - 2024-10-18
+- mod: `IUIManager.GetCustomLayer` to `IUIManager.GetCustomLayerCanvas`, `AbstractUIPanel.SetSortName` to `AbstractUIPanel.SetOrderInLayer`
+- add: `AbstractUIPanel.SetSortingLayer`
 ## [1.8.2] - 2024-10-17
 -mod: improve `PanelCreator` Editor
 ## [1.8.0] - 2024-10-17